+++ rebar.config
@@ -8,11 +8,18 @@
 {port_env, [
 	    {"CFLAGS", "$CFLAGS -Wall -Wextra -Wswitch-default -Wswitch-enum -fno-common -I ../dthread/include"},
 	    {"(darwin)", "CFLAGS", "$CFLAGS -ObjC "},
-	    {"(darwin)", "LDFLAGS", "$LDFLAGS -framework IOBluetooth -framework CoreFoundation -framework Foundation ../dthread/c_src/dlog.o"}
+	    {"(darwin)", "LDFLAGS", "$LDFLAGS -framework IOBluetooth -framework CoreFoundation -framework Foundation ../dthread/c_src/dlog.o"},
+
+	    {"(linux)", "LDFLAGS", "$LDFLAGS -lbluetooth ../dthread/c_src/dlog.o"}
 	    ]}.
 
 %% first version is a regular port program
 {port_specs, [
-	      {"(darwin)", "priv/bt", [ "c_src/bt_macos_drv.c" ]},
-	      {"(linux)", "priv/bt",  [ "c_src/bt_linux_drv.c" ]}
+	      {"(darwin)", "priv/bt", [ "c_src/bt_macos_drv.c",
+					"c_src/bt_sub.c"
+				      ]},
+	      {"(linux)", "priv/bt",  [ "c_src/bt_linux_drv.c",
+					"c_src/bt_sub.c"
+				      ]}
 	     ]}.
+
